Revert "Merge pull request #554 from kinopio-club/test-vue-revert"
This reverts commit 1de405d32f67dff320fd20a1d91c16e5d119eb12, reversing
changes made to be2725c3995fcc8086a86f412104d12b804c41a4.
