+++ package.json
@@ -24,15 +24,15 @@
     "shader": "^1.0.0",
     "smartquotes": "^2.3.2",
     "uuid": "^9.0.0",
-    "vue": "^3.4",
-    "vue-router": "^4.0.11",
+    "vue": "^3.5",
+    "vue-router": "^4.4.3",
     "vuex": "^4.0.2"
   },
   "devDependencies": {
     "@types/lodash-es": "^4.17.3",
-    "@vitejs/plugin-vue": "^5.0.0",
     "@vue/compiler-sfc": "^3.2.6",
     "@vue/eslint-config-standard": "^4.0.0",
+    "@vitejs/plugin-vue": "^5.1.3",
     "autoprefixer": "^10.4.2",
     "babel-eslint": "^10.0.1",
     "eslint": "^7.32.0",
@@ -42,7 +42,7 @@
     "pug": "^3.0.3",
     "stylus": "^0.63",
     "stylus-loader": "^3.0.2",
-    "vite": "^5.0.13",
+    "vite": "^5.4.3",
     "vite-plugin-pwa": "^0.17.4",
     "vite-plugin-sitemap": "^0.7.1",
     "yorkie": "^2.0.0"
