+++ src/components/UserDetailsActions.vue
@@ -16,7 +16,6 @@ const store = useStore()
 const dialogElement = ref(null)
 
 onMounted(() => {
-  window.addEventListener('resize', updateDialogHeight)
   store.subscribe(mutation => {
     if (mutation.type === 'triggerCloseChildDialogs') {
       closeDialogs()
@@ -30,7 +29,6 @@ const props = defineProps({
   showExploreSpaces: Boolean
 })
 const state = reactive({
-  dialogHeight: null,
   spacePickerIsVisible: false,
   exploreSpaces: [],
   userSpaces: [],
@@ -46,17 +44,9 @@ const state = reactive({
 
 watch(() => props.visible, (value, prevValue) => {
   if (value) {
-    updateDialogHeight()
     clearUserSpaces()
   }
 })
-const updateDialogHeight = async () => {
-  if (!props.visible) { return }
-  await nextTick()
-  let element = dialogElement.value
-  state.dialogHeight = utils.elementHeight(element)
-}
-
 const closeDialogs = () => {
   state.spacePickerIsVisible = false
 }
