+++ src/components/dialogs/NewBlankDialogTemplate.vue
@@ -1,11 +1,11 @@
 <script setup>
-import { reactive, computed, onMounted, onBeforeUnmount, defineProps, defineEmits, watch, ref, nextTick } from 'vue'
+import { reactive, computed, onMounted, onBeforeUnmount, defineProps, defineEmits, watch, useTemplateRef, nextTick } from 'vue'
 import { useStore } from 'vuex'
 
 import utils from '@/utils.js'
 const store = useStore()
 
-const dialogElement = ref(null)
+const dialogElement = useTemplateRef('dialogElement')
 
 onMounted(() => {
   window.addEventListener('resize', updateDialogHeight)
