+++ src/components/dialogs/UserDetails.vue
@@ -1,5 +1,5 @@
 <script setup>
-import { reactive, computed, onMounted, onUnmounted, defineProps, defineEmits, watch, ref, nextTick } from 'vue'
+import { reactive, computed, onMounted, onUnmounted, defineProps, defineEmits, watch, useTemplateRef, nextTick } from 'vue'
 import { useStore } from 'vuex'
 
 import UserDetailsInfo from '@/components/UserDetailsInfo.vue'
@@ -8,9 +8,11 @@ import utils from '@/utils.js'
 
 const store = useStore()
 
-const dialogElement = ref(null)
+const dialogElement = useTemplateRef('dialogElement')
 
 onMounted(() => {
+  window.addEventListener('resize', updateDialogHeight)
+  window.addEventListener('scroll', updateDialogHeight)
   store.subscribe((mutation, state) => {
     if (mutation.type === 'triggerScrollUserDetailsIntoView' && visible.value) {
       scrollUserDetailsIntoView()
@@ -18,7 +20,23 @@ onMounted(() => {
   })
 })
 
+const state = reactive({
+  dialogHeight: null
+})
+
 const visible = computed(() => store.state.userDetailsIsVisible)
+watch(() => visible.value, (value, prevValue) => {
+  if (value) {
+    updateDialogHeight()
+  }
+})
+const updateDialogHeight = async () => {
+  if (!visible.value) { return }
+  await nextTick()
+  let element = dialogElement.value
+  state.dialogHeight = utils.elementHeight(element)
+}
+
 const user = computed(() => store.state.userDetailsUser)
 const position = computed(() => store.state.userDetailsPosition)
 
@@ -38,7 +56,8 @@ const styles = computed(() => {
   const styles = {
     transform: `scale(${zoom})`,
     left: x + 'px',
-    top: y + 'px'
+    top: y + 'px',
+    maxHeight: state.dialogHeight + 'px'
   }
   if (transformOriginIsTopRight) {
     styles.transformOrigin = 'top right'
@@ -65,4 +84,5 @@ dialog.user-details
   cursor initial
   top calc(100% - 8px)
   position absolute
+  overflow auto
 </style>
